feat: add configurable HTML-ID prefix for SVG codes
This adds an option for callers to provide an HTML-ID prefix for SVG
QR-codes, in order to make them unique and enable using this library for
multiple QR-codes on a single page.

Further it adds a .node-version file which can be used by nodenv or just
manual node version selection, to specify a node version that is able to
install, build and test this code base without adding any changes.

Closes #210.